refactor: change tree related files to use composition api (#1663)
* refactor: change tree related components to use composition api


* refactor: do not use vm in tree component


* refactor: remove utils.ts


* refactor: remove unnecessary method from Tree.vue


* chore: add comments


* refactor: split into several files


* fix: update DragEventService


---------
