--- packages/mirinae/.eslintignore
@@ -1,4 +1,0 @@
-# he-tree-vue
-src/data-display/tree/he-tree-vue/components/Tree.vue
-src/data-display/tree/he-tree-vue/plugins/*
-src/data-display/tree/he-tree-vue/helpers.ts
