+++ packages/mirinae/src/data-display/tree/PTree.vue
@@ -101,14 +101,11 @@ import {
 import { unionBy } from 'lodash';
 import { focus } from 'vue-focus';
 
-import type { Store } from '@/data-display/tree/he-tree-vue/plugins/draggable/draggable-types';
-import type {
-    CloneTreeDataOptions, WalkTreeDataCallback,
-} from '@/data-display/tree/he-tree-vue/types';
+import type { Store } from '@/data-display/tree/he-tree-vue/libs/draggable/types';
 import { getDefaultNode } from '@/data-display/tree/helper';
 import type {
-    TreeNode,
     Tree,
+    TreeNode,
     Predicate,
     TreeItem,
     ToggleOptions,
@@ -117,19 +114,17 @@ import type {
     DragOptions,
     DataGetter,
     DataSetter,
-    GetClassNames, DataFetcher,
+    GetClassNames, DataFetcher, CloneTreeDataOptions,
 } from '@/data-display/tree/type';
 import PSpinner from '@/feedbacks/loading/spinner/PSpinner.vue';
 import PI from '@/foundation/icons/PI.vue';
 import PTextInput from '@/inputs/input/text-input/PTextInput.vue';
 
 import OriginTree from './he-tree-vue/components/Tree.vue';
-import DraggablePlugin from './he-tree-vue/plugins/draggable/Draggable.vue';
-import FoldPlugin from './he-tree-vue/plugins/fold';
-import { walkTreeData, cloneTreeData } from './he-tree-vue/utils';
+import type { WalkTreeDataCallback } from './he-tree-vue/tree-data';
+import { walkTreeData, cloneTreeData } from './he-tree-vue/tree-data';
 
 
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface Props {
     toggleOptions: ToggleOptions;
     selectOptions: SelectOptions;
@@ -142,17 +137,15 @@ interface Props {
     getClassNames: GetClassNames;
 }
 
+
+
 export default defineComponent<Props>({
     name: 'PTree',
     components: {
         PSpinner,
         PTextInput,
         PI,
-        Tree: {
-            name: 'Tree',
-            extends: OriginTree,
-            mixins: [FoldPlugin, DraggablePlugin],
-        },
+        Tree: OriginTree,
     },
     directives: { focus },
     props: {
@@ -288,9 +281,9 @@ export default defineComponent<Props>({
             setSelectItem(node, path, value);
         };
 
-        const onDragStart = (tree: any, e: Store) => {
+        const onDragStart = (e: Store) => {
             const validator = props.dragOptions.startValidator;
-            const parent = e.startPath ? tree.getNodeParentByPath(e.startPath) as TreeNode : null;
+            const parent = e.startPath ? state.treeRef.getNodeParentByPath(e.startPath) as TreeNode : null;
 
             if (validator && !validator(e.dragNode as TreeNode, parent)) return false;
 
@@ -298,10 +291,10 @@ export default defineComponent<Props>({
             return true;
         };
 
-        const onDragEnd = (tree: any, e: Store) => {
+        const onDragEnd = (e: Store) => {
             state.dragTargetParentPath = null;
-            const parent = e.targetPath ? tree.getNodeParentByPath(e.targetPath) as TreeNode : null;
-            const oldParent = e.startPath ? tree.getNodeParentByPath(e.startPath) as TreeNode : null;
+            const parent = e.targetPath ? state.treeRef.getNodeParentByPath(e.targetPath) as TreeNode : null;
+            const oldParent = e.startPath ? state.treeRef.getNodeParentByPath(e.startPath) as TreeNode : null;
 
             const validator = props.dragOptions.endValidator;
 
@@ -330,32 +323,32 @@ export default defineComponent<Props>({
             let oldParent: TreeNode|null = null;
             if (e.startPath) {
                 try {
-                    oldParent = e.startTree?.getNodeParentByPath(e.startPath) as TreeNode;
+                    oldParent = state.treeRef.getNodeParentByPath(e.startPath) as TreeNode;
                 } catch (error) {}
             }
 
-            const parent = e.targetTree?.getNodeParentByPath(targetPath) as TreeNode;
+            const parent = state.treeRef.getNodeParentByPath(targetPath) as TreeNode;
 
             emit('drop', e.dragNode, oldParent, parent, rollback);
         };
 
-        const eachDraggable = (path: number[], tree: any, e: Store) => {
+        const eachDraggable = (path: number[], e: Store) => {
             const dragValidator = props.dragOptions.dragValidator;
             if (dragValidator
-                && !dragValidator(e.dragNode as TreeNode, path ? tree.getNodeParentByPath(path) as TreeNode : undefined)) {
+                && !dragValidator(e.dragNode as TreeNode, path ? state.treeRef.getNodeParentByPath(path) as TreeNode : undefined)) {
                 return false;
             }
             return true;
         };
-        const eachDroppable = (parentPath: number[], tree: any, e: Store) => {
+        const eachDroppable = (parentPath: number[], e: Store) => {
             const dropValidator = props.dragOptions.dropValidator;
 
             let parent: null|TreeNode = null;
             if (parentPath.length > 0) {
-                parent = tree.getNodeByPath(parentPath) as TreeNode;
+                parent = state.treeRef.getNodeByPath(parentPath) as TreeNode;
             }
 
-            const oldParent = e.startPath ? tree.getNodeParentByPath(e.startPath) : null;
+            const oldParent = e.startPath ? state.treeRef.getNodeParentByPath(e.startPath) : null;
 
             state.dragTargetParentPath = parentPath;
