+++ packages/mirinae/src/data-display/tree/he-tree-vue/components/Tree.vue
@@ -1,183 +1,181 @@
 <template>
-    <div class="he-tree" :class="{'he-tree--rtl': rtl}" :data-tree-id="treeId">
-        <children-list :nodes="rootNode.children"
-                       :indent="indent"
-                       :parent="rootNode"
+    <div ref="treeRef"
+         class="he-tree"
+         :class="{'he-tree--rtl': props.rtl}"
+         :data-tree-id="state.treeId"
+    >
+        <children-list :nodes="state.rootNode.children"
+                       :indent="props.indent"
+                       :parent="state.rootNode"
                        :parent-path="[]"
-                       :root-node="rootNode"
-                       :rtl="rtl"
-                       :selected-paths="selectedPaths"
+                       :root-node="state.rootNode"
+                       :rtl="props.rtl"
+                       :selected-paths="props.selectedPaths"
         >
             <template #default="scope">
-              <slot v-bind="scope" />
+                <slot v-bind="scope" />
             </template>
         </children-list>
     </div>
 </template>
 
-<script lang="ts">
-// @ts-nocheck
-import * as ut from '../utils';
-import { defineComponent } from 'vue';
+<script setup lang="ts">
 import {
-  arrayRemove,
-  joinFunctionsByNext,
-  randString,
-  TreeData, updatablePropsEvenUnbound
-} from "../helpers";
+    reactive, toRefs,
+    computed, watch, ref,
+    onBeforeMount, defineExpose, toRef,
+    defineAsyncComponent,
+} from 'vue';
 
-const trees = {};
-const ChildrenList = () => ({
-  component: import('./ChildrenList.vue') as any,
-})
+import { useDraggable } from '@/data-display/tree/he-tree-vue/composables/use-draggable';
+import { useFold } from '@/data-display/tree/he-tree-vue/composables/use-fold';
+import type { Store } from '@/data-display/tree/he-tree-vue/libs/draggable/types';
+import {
+    TreeData, walkTreeData as _walkTreeData, cloneTreeData as _cloneTreeData, getPureTreeData as _getPureTreeData,
+} from '@/data-display/tree/he-tree-vue/tree-data';
+
+const ChildrenList = defineAsyncComponent(() => import('./ChildrenList.vue'));
+
+interface Props {
+    value?: any[];
+    selectedPaths?: number[];
+    // tree draggable options
+    indent?: number;
+    // draggable helper options
+    rtl?: boolean;
+    draggable?: boolean;
+    droppable?: boolean;
+    eachDraggable?: (path: number[], e: Store) => boolean;
+    eachDroppable?: (path: number[], e: Store) => boolean;
+    ondragstart?: (e: Store) => boolean;
+    ondragend?: (e: Store) => boolean;
+    unfoldWhenDragover?: boolean;
+    unfoldWhenDragoverDelay?: number;
+    // fold
+    foldingTransitionNames?: string;
+    foldingTransition?: boolean;
+    foldAllAfterMounted?: boolean;
+}
+const props = withDefaults(defineProps<Props>(), {
+    value: () => [],
+    indent: 20,
+    rtl: false,
+    selectedPaths: () => [],
+    // draggable
+    draggable: true,
+    droppable: true,
+    eachDraggable: undefined,
+    eachDroppable: undefined,
+    ondragstart: undefined,
+    ondragend: undefined,
+    unfoldWhenDragover: true,
+    unfoldWhenDragoverDelay: 30,
+    // fold
+    foldingTransitionNames: undefined,
+    foldingTransition: undefined,
+    foldAllAfterMounted: undefined,
+});
+
+// TODO: add events from composables
+const emit = defineEmits<{(event: 'input', value: any[]): void;
+    (event: 'update:value', value: any[]): void;
+    (event: 'update:root-node', value: any[]): void;
+}>();
+
+const state = reactive({
+    rootNode: {} as any,
+    treeId: Math.floor(Math.random() * Date.now()),
+    proxyValue: [] as any[],
+    treeData: computed({
+        get() {
+            return state.proxyValue;
+        },
+        set(value) {
+            emit('update:value', value);
+            emit('input', value);
+            state.proxyValue = value;
+        },
+    }),
+});
 
 
-export default defineComponent({
-    name: 'Tree',
-    components: {
-        ChildrenList
-    },
-    props: {
-        value: {
-          type: Array,
-          default: () => []
-        },
-        indent: {
-          type: Number,
-          default: 20
-        },
-        rtl: {
-          type: Boolean,
-          default: false
-        },
-        rootNode: {
-          type: Object,
-          default: () => ({})
-        },
-        selectedPaths: {
-            type: Array,
-            default: () => []
-        }
-    },
-    mixins: [
-      updatablePropsEvenUnbound({
-        value: { $localName: 'treeData', required: true },
-      }),
-    ],
-    data() {
-        return {
-            trees,
-            treeClass: '',
-            treeId: randString(),
-            _hooks: {} as Record<string, {(...args: any[]): any}[]>
-        };
-    },
-    watch: {
-        treeData: {
-            immediate: true,
-            handler(treeData) {
-                this._TreeDataHelper = new TreeData(this.treeData);
-            },
-        },
-    },
-    methods: {
-      //
-      // get hooks in this._hooks, without which in props
-      _getNonPropHooksByName(name) {
-        if (this._hooks) {
-          return this._hooks[name]
-        }
-      },
-      addHook(name, func) {
-        if (!this._getNonPropHooksByName(name)) {
-          if (!this._hooks) {
-            this._hooks = {}
-          }
-          if (!this._hooks[name]) {
-            this._hooks[name] = []
-          }
-        }
-        this._hooks[name].push(func)
-      },
-      removeHook(name, func) {
-        const hooks = this._getNonPropHooksByName(name)
-        if (hooks) {
-          arrayRemove(hooks, func)
-        }
-      },
-      hasHook(name) {
-        return this._getNonPropHooksByName(name) || this[name]
-      },
-      executeHook(name, hookArgs: any[]) {
-        let hooks = this._getNonPropHooksByName(name)
-        hooks = hooks ? hooks.slice() : []
-        const func = this[name]
-        if (func && typeof func === 'function') {
-          hooks.push((next, ...args) => {
-            return func(...args)
-          })
-        }
-        return joinFunctionsByNext(hooks)(...hookArgs)
-      },
+watch(() => props.value, (value) => {
+    state.proxyValue = value;
+}, { immediate: true });
 
-      //
-        iteratePath(path, opt) {
-            // @ts-ignore
-            return this._TreeDataHelper.iteratePath(path, opt);
-        },
-        getTreeVmByTreeEl(treeEl) {
-            return this.trees[treeEl.getAttribute('data-tree-id')];
-        },
-        getAllNodesByPath(path) {
-            // @ts-ignore
-            return this._TreeDataHelper.getAllNodes(path);
-        },
-        getNodeByPath(path) {
-            // @ts-ignore
-            return this._TreeDataHelper.getNode(path);
-        },
-        getPathByBranchEl(branchEl) {
-            return branchEl.getAttribute('data-tree-node-path').split(',').map(v => parseInt(v));
-        },
-        getBranchElByPath(path) {
-            return this.$el.querySelector(`[data-tree-node-path='${path.join(',')}']`);
-        },
-        getNodeByBranchEl(branchEl) {
-            return this.getNodeByPath(this.getPathByBranchEl(branchEl));
-        },
-        getNodeParentByPath(path) {
-            // @ts-ignore
-            return this._TreeDataHelper.getNodeParent(path);
-        },
-        removeNodeByPath(path) {
-            // @ts-ignore
-            return this._TreeDataHelper.removeNode(path);
-        },
-        walkTreeData(handler, opt) {
-            return ut.walkTreeData(this.treeData, handler, opt);
-        },
-        cloneTreeData(opt) {
-            return ut.cloneTreeData(this.treeData, opt);
-        },
-        // return cloned new tree data without property witch starts with `$`
-        getPureTreeData() {
-            return ut.getPureTreeData(this.treeData);
-        },
-    },
-    created() {
-        //
-        const updateRootNode = () => { this.$set(this.rootNode, 'children', this.treeData); };
-        this.$watch('rootNode', updateRootNode, { immediate: true });
-        this.$watch('treeData', updateRootNode, { immediate: true });
-    },
-    mounted() {
-        //
-        this.treeId = randString();
-        this.$set(this.trees, this.treeId, this);
-        this.$once('hook:beforeDestroy', () => {
-            this.$delete(this.trees, this.treeId);
-        });
-    },
+const treeRef = ref<HTMLElement|null>(null);
+
+let treeDataHelper: TreeData<any>;
+
+watch(() => state.treeData, (treeData) => {
+    treeDataHelper = new TreeData(treeData);
+}, { immediate: true });
+
+const iteratePath = (path: number[], opt) => treeDataHelper.iteratePath(path, opt);
+const getAllNodesByPath = (path: number[]) => treeDataHelper.getAllNodes(path);
+const getNodeByPath = (path: number[]) => treeDataHelper.getNode(path);
+const getBranchElByPath = (path: number[]): Element|null => {
+    if (treeRef.value) {
+        return treeRef.value.querySelector(`[data-tree-node-path='${path.join(',')}']`);
+    }
+    return null;
+};
+const getNodeByBranchEl = (branchEl: HTMLElement) => getNodeByPath(draggableMethods.getPathByBranchEl(branchEl));
+const getNodeParentByPath = (path: number[]) => treeDataHelper.getNodeParent(path);
+const removeNodeByPath = (path: number[]) => treeDataHelper.removeNode(path);
+const walkTreeData = (handler, opt) => _walkTreeData(state.treeData, handler, opt);
+const cloneTreeData = (opt) => _cloneTreeData(state.treeData, opt);
+// return cloned new tree data without property witch starts with `$`
+const getPureTreeData = () => _getPureTreeData(state.treeData);
+
+const methods = {
+    iteratePath,
+    getAllNodesByPath,
+    getNodeByPath,
+    getBranchElByPath,
+    getNodeByBranchEl,
+    getNodeParentByPath,
+    removeNodeByPath,
+    walkTreeData,
+    cloneTreeData,
+    getPureTreeData,
+};
+
+onBeforeMount(() => {
+    const updateRootNode = () => {
+        state.rootNode.children = state.treeData;
+        emit('update:root-node', state.rootNode);
+    };
+    watch(() => state.rootNode, updateRootNode, { immediate: true });
+    watch(() => state.treeData, updateRootNode, { immediate: true });
+});
+
+const { methods: foldMethods } = useFold(props, emit, { walkTreeData });
+
+const { treesStore, methods: draggableMethods } = useDraggable(props, emit, {
+    treeRef,
+    rootNode: toRef(state, 'rootNode'),
+    treeData: toRef(state, 'treeData'),
+    getAllNodesByPath,
+    getNodeByPath,
+    getNodeByBranchEl,
+    getBranchElByPath,
+    iteratePath,
+    unfold: foldMethods.unfold,
+});
+
+
+
+const allMethods = {
+    ...methods,
+    ...foldMethods,
+    ...draggableMethods,
+};
+
+defineExpose({
+    ...toRefs(state),
+    treesStore,
+    ...allMethods,
 });
 
 </script>
